requirements.txt
click>=8.0
PyYAML>=6.0
pandas>=1.5
openpyxl>=3.0
